modified extract methods - created prodivers csv - first test

+++ webscrape_bids.py
@@ -46,28 +46,37 @@ class BidScrape():
     def extract(self):
         time.sleep(5)
         selector = self.selector
-        contract_info = selector("m", "table[id*='DetalleImputacionAdjudicacion'] td") 
+        #MAIN DICT
         data_main = {         
-            "bid_code": selector("s", "span[id*='NumeroProceso']"), 
+            "code": selector("s", "span[id*='NumeroProceso']"), 
             "name": selector("s", "span[id*='NombreProceso']"), 
             "process": selector("s", "span[id*='ProcedimientoSeleccion']"), 
             "stage": selector("s", "span[id*='Etapa']"), 
             "validity": selector("s", "span[id*='MantenimientoOferta']"), 
             "duration": selector("s", "span[id*='DuracionContrato']"), 
-            "opening": selector("s", "table[id*='ActasApertura'] p"), 
-            "awarded_bidder": contract_info[1].text, 
-            "awarded_bidder_id": contract_info[2].text, 
-            "amount": contract_info[6].text, 
-            "currency": contract_info[7].text      
+            "opening": selector("s", "table[id*='ActasApertura'] p"),     
             } 
-        products = [product.text for product in selector("m", "table[id*='gvLineaPliego']  span[id*='lblDescripcion']")]
-        products_qty = [qty.text for qty in selector("m", "table[id*='gvLineaPliego']  span[id*='lblCantidad']")]
-        data_products = {"bid_code": [], "product": [], "qty": []}
-        for prod in range(len(products)):
+        
+        #PROVIDERS DICT
+        data_providers = {"bid_code": [], "name": [], "tin": [], "po_number": [], "po_amount": [], "currency": []}
+        for provider_rows in selector("m", "table[id*='DetalleImputacionAdjudicacion'] tr")[1:]:
+            row = provider_rows.find_elements_by_tag_name("td")
+            data_providers["bid_code"].append(selector("s", "span[id*='NumeroProceso']"))
+            data_providers["name"].append(row[1].text)
+            data_providers["tin"].append(row[2].text)
+            data_providers["po_number"].append(row[0].text)
+            data_providers["po_amount"].append(row[6].text)
+            data_providers["currency"].append(row[7].text)
+        
+        #PRODUCTS DICT
+        data_products = {"bid_code": [], "description": [], "qty": []}
+        for product_rows in selector("m", "table[id*='DetalleProductos_gvLineaPliego'] tr")[1:]:
+            row = product_rows.find_elements_by_tag_name("td")  
             data_products["bid_code"].append(selector("s", "span[id*='NumeroProceso']"))
-            data_products["product"].append(products[prod])
-            data_products["qty"].append(products_qty[prod])
-        return data_main, data_products
+            data_products["description"].append(row[3].text)
+            data_products["qty"].append(row[4].text)
+    
+        return data_main, data_providers, data_products
             
     def exit_process(self):
         self.driver.execute_script("arguments[0].click();", WebDriverWait(self.driver, 30).until(EC.element_to_be_clickable((By.XPATH,  f"//*[@id='ctl00_CPH1_lnkVolver']"))))
@@ -104,8 +113,10 @@ class BidScrape():
         driver = self.driver
         if self.iteration == 0:
             self.first_iteration()
-        data_main = {"bid_code": [], "name": [], "process": [], "stage": [], "validity": [], "duration": [], "opening": [], "awarded_bidder": [], "awarded_bidder_id": [], "amount": [], "currency": []}
-        data_products = {"bid_code": [], "product": [], "qty": []}
+        data_main = {"code": [], "name": [], "process": [], "stage": [], "validity": [], "duration": [], "opening": []}
+        data_providers = {"bid_code": [], "name": [], "tin": [], "po_number": [], "po_amount": [], "currency": []}
+        data_products = {"bid_code": [], "description": [], "qty": []}
+        
         for row in range(1, 11):
             f_row = format(row+1, "02d")
             self.enter_process(f_row)
@@ -113,10 +124,16 @@ class BidScrape():
             self.exit_process()
             print("Row " + str(row) + " scraped")
             self.row_counter += 1
-            for keys in data_main:
+            
+            for keys in extracted_data[0]: #MAIN DICT
                 data_main[keys].append(extracted_data[0][keys])
-            for keys in data_products:
-                data_products[keys].append(extracted_data[1][keys])
+            for key in extracted_data[1]: #PROVIDERS DICT
+                for value in extracted_data[1][key]:
+                    data_providers[key].append(value)     
+            for key in extracted_data[2]: #PRODUCTS DICT
+                for value in extracted_data[2][key]:
+                    data_products[key].append(value)
+                    
         if self.tab_counter % 10 == 0 and self.tab_counter != 0:
             self.page_jump([11, 12])
             self.tab_counter += 1
@@ -125,7 +142,7 @@ class BidScrape():
             self.tab_counter += 1
             self.tab_jump(self.tab_counter)
         self.iteration += 1
-        return data_main, data_products
+        return data_main, data_providers, data_products
         
 
 #COUNTERS PULL
@@ -146,24 +163,24 @@ compras_ar.query_search()
 
 #CSV LOAD
 path_main = fr"local_repo\bids\bids_report.csv"
+path_providers = fr"local_repo\bids\providers_report.csv"
 path_products = fr"local_repo\bids\products_report.csv"
 
 for n in range(1):
-    main_data, product_data = compras_ar.scrape()
+    data_main, data_providers, data_products = compras_ar.scrape()
     
-    df_main = pd.DataFrame(main_data)
-    df_products = pd.DataFrame(product_data)
+    df_main = pd.DataFrame(data_main)
+    df_providers = pd.DataFrame(data_providers)
+    df_products = pd.DataFrame(data_products)
     
-    if path.exists(path_main):
-        df_main.to_csv(path_main, mode="a", index=False, header=False)
-    else:
+    if not path.exists(path_main) and not path.exists(path_providers) and not path.exists(path_products):
         df_main.to_csv(path_main, index=False)
-        
-    if path.exists(path_products):
-        df_products.to_csv(path_products, mode="a", index=False, header=False)
-    else:
+        df_providers.to_csv(path_providers, index=False)
         df_products.to_csv(path_products, index=False)
-        
+    else:
+        df_main.to_csv(path_main, mode="a", index=False, header=False)
+        df_providers.to_csv(path_providers, mode="a", index=False, header=False)
+        df_products.to_csv(path_products, mode="a", index=False, header=False)
 
 compras_ar.driver.quit()
 
@@ -177,16 +194,16 @@ with open(path_counters, "w") as write_file:
 t1_stop = perf_counter() 
 time_info = f"Elapsed time: {str(t1_start)}(Start), {str(t1_stop)}(Stop) \nElapsed time during the whole program in seconds: {str(t1_stop-t1_start)}"
 timestamp = f"Timestamp: {datetime.now()}"
-pages = f"Pages completed: {compras_ar.page_counter - page_counter_current}\nCurrent page: {compras_ar.page_counter}"
-tabs = f"Tabs completed: {compras_ar.tab_counter - tab_counter_current}\nCurrent tab: {compras_ar.tab_counter}"
+pages = f"Pages completed: {compras_ar.page_counter - page_counter_current - 1}\nCurrent page: {compras_ar.page_counter}"
+tabs = f"Tabs completed: {compras_ar.tab_counter - tab_counter_current - 1}\nCurrent tab: {compras_ar.tab_counter}"
 rows = f"Rows completed: {compras_ar.row_counter}"
 
-log = time_info + "\n" + timestamp + "\n" + pages + "\n" + tabs + "\n" + rows
+log = time_info + "\n" + timestamp + "\n" + pages + "\n" + tabs + "\n" + rows + "\n" + "\n------------------------\n"
 print(log)
 
 if path.exists(path_counters):
-    with open(fr"local_repo\bids\bidslog.txt", "a+") as log_file:
-        log_file.write("\n------------------------\n")
+    with open(fr"local_repo\bids\bidslog.txt", "a+") as file:
+        file.write(log)       
 else:
     with open(fr"local_repo\bids\bidslog.txt", "w") as file:
         file.write(log)
